Set limits and input types for the component Form Feature/54.admin-page-form
